fix bug cgi

+++ srcs/cgi/Cgi.cpp
@@ -6,7 +6,7 @@
 /*   By: jusilanc <jusilanc@s19.be>                 +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/11 16:28:09 by jusilanc          #+#    #+#             */
-/*   Updated: 2023/11/21 17:25:49 by jusilanc         ###   ########.fr       */
+/*   Updated: 2023/11/21 17:35:15 by jusilanc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -229,7 +229,7 @@ const std::string& Cgi::run()
 		close(fdOut[1]);
 		dup2(fdIn[1], STDIN_FILENO);
 		dup2(fdOut[0], STDOUT_FILENO);
-		int retW = 0;
+		size_t retW = 0;
 		if (_method == "POST" || _method == "DELETE")
 		{
 			std::stringstream ss;
